build: reset project version
For now not tying it to minecraft-data, at least until API is stable.

Some breaking changes to this library API are planned, and it's impossible
to follow SemVer and have a meaningful version number for this library if
it simply clones the minecraft-data's version it ships (and minecraft-data
itself does follow Semver).

For example, how to version distinct releases that ships the same version
of minecraft-data?

A possible solution is to use 1.2.3+mcd-4.5.6 as version pattern, as it
would indicate the minecraft-data version while still being fully compliant
with SemVer and able to communicate any API changes, new features and bug
fixes in _this_ library by its version number.

In the future, cloning the version number might be considered again if the
API gets stable enough and no major bugs are found, so any fix can wait until
the next minecraft-data version.

Meanwhile, 0.1 it is.